requirements.txt
line-bot-sdk
flask
gunicorn
openai
langchain
FinMind
uvicorn
fastapi